(PC-16306)[PRO] feat: remove ff OFFER_FORM_SUMMARY_PAGE
