+++ pro/src/components/pages/Offers/Offer/OfferStatus/__specs__/StatusLabel.spec.tsx
@@ -2,7 +2,6 @@ import '@testing-library/jest-dom'
 
 import { render, screen } from '@testing-library/react'
 import React from 'react'
-import { Provider } from 'react-redux'
 
 import {
   OFFER_STATUS_ACTIVE,
@@ -12,57 +11,36 @@ import {
   OFFER_STATUS_REJECTED,
   OFFER_STATUS_SOLD_OUT,
 } from 'core/Offers/constants'
-import { configureTestStore } from 'store/testUtils'
 
 import StatusLabel from '../StatusLabel'
 
-const renderStatusLabel = (status: string, store: any) => {
-  return render(
-    <Provider store={configureTestStore(store)}>
-      <StatusLabel status={status} />
-    </Provider>
-  )
+const renderStatusLabel = (status: string) => {
+  return render(<StatusLabel status={status} />)
 }
 
 describe('StatusLabel', () => {
-  let store = {}
   it('should display "expirée" if offer is expired', () => {
-    renderStatusLabel(OFFER_STATUS_EXPIRED, store)
+    renderStatusLabel(OFFER_STATUS_EXPIRED)
     expect(screen.getByText('expirée')).toBeInTheDocument()
   })
-  it('should display "expirée" if offer is active', () => {
-    renderStatusLabel(OFFER_STATUS_ACTIVE, store)
-    expect(screen.getByText('active')).toBeInTheDocument()
+  it('should display "publiée" if offer is active', () => {
+    renderStatusLabel(OFFER_STATUS_ACTIVE)
+    expect(screen.getByText('publiée')).toBeInTheDocument()
   })
-  it('should display "expirée" if offer is inactive', () => {
-    renderStatusLabel(OFFER_STATUS_INACTIVE, store)
+  it('should display "désactivée" if offer is inactive', () => {
+    renderStatusLabel(OFFER_STATUS_INACTIVE)
     expect(screen.getByText('désactivée')).toBeInTheDocument()
   })
-  it('should display "expirée" if offer is pending', () => {
-    renderStatusLabel(OFFER_STATUS_PENDING, store)
+  it('should display "en attente" if offer is pending', () => {
+    renderStatusLabel(OFFER_STATUS_PENDING)
     expect(screen.getByText('en attente')).toBeInTheDocument()
   })
-  it('should display "expirée" if offer is rejected', () => {
-    renderStatusLabel(OFFER_STATUS_REJECTED, store)
+  it('should display "refusée" if offer is rejected', () => {
+    renderStatusLabel(OFFER_STATUS_REJECTED)
     expect(screen.getByText('refusée')).toBeInTheDocument()
   })
-  it('should display "expirée" if offer is sold out', () => {
-    renderStatusLabel(OFFER_STATUS_SOLD_OUT, store)
+  it('should display "épuisée" if offer is sold out', () => {
+    renderStatusLabel(OFFER_STATUS_SOLD_OUT)
     expect(screen.getByText('épuisée')).toBeInTheDocument()
-  })
-
-  it('should display "publiée" if offer is active with FF', () => {
-    store = {
-      features: {
-        list: [
-          {
-            isActive: true,
-            nameKey: 'OFFER_FORM_SUMMARY_PAGE',
-          },
-        ],
-      },
-    }
-    renderStatusLabel(OFFER_STATUS_ACTIVE, store)
-    expect(screen.getByText('publiée')).toBeInTheDocument()
   })
 })
